pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "butterfly"
version = "2.0.0"
description = "A minimal Python Agent library"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0"]

[project.scripts]
butterfly               = "ui.cli.main:main"
butterfly-server        = "butterfly.runtime.server:main"
butterfly-web           = "ui.web.app:main"

[tool.hatch.build.targets.wheel]
packages = ["butterfly", "ui"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
